requirements.txt
django
djangorestframework
pillow
django-cors-headers
whitenoise
gunicorn
six
django-utils-six
Django-Verify-Email
psycopg2
dj-database-url